update user structure

+++ models/models.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type User struct {
-	ID int64 `json:"id"`
-	Name string `json:"name"`
+	ID         int       `json:"id"`
+	Name       string    `json:"name"`
 	DateJoined time.Time `json:"date_joined"`
 }
